perf: refactor
