+++ src/constants/_providers.ts
@@ -1,0 +1,3 @@
+export const PROVIDERS = {
+  GITHUB: "github",
+} as const;
