+++ src/constants/_texts.ts
@@ -1,23 +1,5 @@
 import { ResultSectionState } from "@/sections/result";
 
-export const CONSTANTS = {
-  COLORS: {
-    DEFAULT: "383838",
-    TEXT: "383838",
-    BACKGROUND: "ffffff",
-  },
-  ID: {
-    TEXT: "textColor",
-    BACKGROUND: "backgroundColor",
-  },
-};
-
-export const contrastColors = {
-  good: "bg-green-600/30 text-green-900",
-  warning: "bg-yellow-600/30 text-yellow-900",
-  error: "bg-red-600/30 text-red-900",
-};
-
 interface TextSizes {
   title: string;
   assessment: keyof ResultSectionState;
@@ -38,7 +20,3 @@ export const textSizes: TextSizes[] = [
       "rounded-xl lg:rounded-br-xl lg:rounded-l-none lg:rounded-tr-none",
   },
 ];
-
-export const PROVIDERS = {
-  GITHUB: "github",
-};
