+++ tailwind.config.ts
@@ -6,7 +6,7 @@ const config: Config = {
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/sections/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/config/constants.ts",
+    "./src/constants/*.ts",
   ],
   theme: {
     extend: {
